chore(worker): remove legacy queue run loop
