+++ internal/worker/runtask_integration_test.go
@@ -8,9 +8,7 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/xrf9268-hue/aiops-platform/internal/task"
 	worker "github.com/xrf9268-hue/aiops-platform/internal/worker"
@@ -681,153 +679,6 @@ func indexOfEvent(ev *fakeEmitter, kind string) int {
 	return -1
 }
 
-// fakeRunStore implements the structural contract Run requires from its
-// store (Claim/Complete + EventEmitter + failingStore) so Run can be
-// driven without Postgres. It hands out a single task, then synthesises
-// ctx cancellation via the onClaimed hook so the worker loop exits
-// deterministically after the failure path runs.
-type fakeRunStore struct {
-	task       *task.Task
-	failResult bool
-	failErr    error
-
-	mu        sync.Mutex
-	claimed   bool
-	events    []recordedEvent
-	failCalls int
-	onClaimed func()
-}
-
-func (s *fakeRunStore) Claim(ctx context.Context) (*task.Task, error) {
-	s.mu.Lock()
-	already := s.claimed
-	s.claimed = true
-	s.mu.Unlock()
-	if already {
-		<-ctx.Done()
-		return nil, ctx.Err()
-	}
-	return s.task, nil
-}
-
-func (s *fakeRunStore) Complete(_ context.Context, _ string) error { return nil }
-
-func (s *fakeRunStore) AddEvent(_ context.Context, _, kind, msg string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.events = append(s.events, recordedEvent{Kind: kind, Message: msg})
-	return nil
-}
-
-func (s *fakeRunStore) AddEventWithPayload(_ context.Context, _, kind, msg string, payload any) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.events = append(s.events, recordedEvent{Kind: kind, Message: msg, Payload: payload})
-	return nil
-}
-
-func (s *fakeRunStore) Fail(_ context.Context, _, _ string) (bool, error) {
-	s.mu.Lock()
-	s.failCalls++
-	cb := s.onClaimed
-	s.mu.Unlock()
-	if cb != nil {
-		cb()
-	}
-	return s.failResult, s.failErr
-}
-
-func (s *fakeRunStore) FailTerminal(_ context.Context, _, _ string) error {
-	s.mu.Lock()
-	s.failCalls++
-	cb := s.onClaimed
-	s.mu.Unlock()
-	if cb != nil {
-		cb()
-	}
-	return s.failErr
-}
-
-func (s *fakeRunStore) FailTimeout(_ context.Context, _, _ string, _ int) (bool, error) {
-	// runTask in these tests fails at PrepareGitWorkspace, which is not a
-	// runner.TimeoutError, so handleTaskFailure always takes the Fail()
-	// path. Implementing FailTimeout here just keeps the interface satisfied.
-	return false, nil
-}
-
-func (s *fakeRunStore) byKind(kind string) []recordedEvent {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	var out []recordedEvent
-	for _, e := range s.events {
-		if e.Kind == kind {
-			out = append(out, e)
-		}
-	}
-	return out
-}
-
 func shellQuoteForTest(s string) string {
 	return "'" + strings.ReplaceAll(s, "'", "'\\''") + "'"
-}
-
-// TestRun_DoesNotWriteTrackerOnFailure pins the SPEC §1 boundary for the
-// failure path: terminality only controls queue retry state. The worker must
-// not construct tracker transitioners or move the linked issue to Rework.
-func TestRun_DoesNotWriteTrackerOnFailure(t *testing.T) {
-	tk := &task.Task{
-		ID:            "tsk_fail",
-		CloneURL:      "file:///definitely-not-a-real-path/spec-boundary",
-		BaseBranch:    "main",
-		WorkBranch:    "ai/tsk_fail",
-		SourceType:    "linear_issue",
-		SourceEventID: "issue-uuid",
-	}
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	store := &fakeRunStore{
-		task:       tk,
-		failResult: true,
-	}
-	store.onClaimed = func() {
-		go func() {
-			time.Sleep(20 * time.Millisecond)
-			cancel()
-		}()
-	}
-
-	cfg := worker.Config{
-		WorkspaceRoot: t.TempDir(),
-		MirrorRoot:    t.TempDir(),
-		Workflow: &workflow.Workflow{
-			Config:         workflow.DefaultConfig(),
-			PromptTemplate: workflow.DefaultPrompt(),
-			Source:         workflow.SourceDefault,
-		},
-	}
-
-	done := make(chan struct{})
-	go func() {
-		worker.Run(ctx, store, cfg)
-		close(done)
-	}()
-	select {
-	case <-done:
-	case <-time.After(5 * time.Second):
-		cancel()
-		t.Fatal("Run did not exit within 5s")
-	}
-
-	if store.failCalls != 1 {
-		t.Fatalf("Fail calls = %d, want 1", store.failCalls)
-	}
-	for _, forbidden := range []string{
-		task.EventTrackerTransition,
-		task.EventTrackerTransitionError,
-		task.EventTrackerComment,
-	} {
-		if got := len(store.byKind(forbidden)); got != 0 {
-			t.Fatalf("worker emitted forbidden tracker event %q %d time(s); events=%#v", forbidden, got, store.events)
-		}
-	}
 }
